feat: 장바구니 상품의 소유주가 맞는지 검증하는 기능 추가

+++ src/main/java/cart/service/CartService.java
@@ -9,6 +9,7 @@ import cart.entity.CartEntity;
 import cart.entity.MemberEntity;
 import cart.entity.product.ProductEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 import java.util.stream.Collectors;
@@ -26,15 +27,16 @@ public class CartService {
         this.memberDao = memberDao;
     }
 
-    public Long putInCart(final Long productId, final MemberAuthDto memberAuthDto) {
-        final ProductEntity product = getProduct(productId);
+    @Transactional(readOnly = true)
+    public List<CartProductResponseDto> findCartItemsForMember(final MemberAuthDto memberAuthDto) {
         final MemberEntity member = getMember(memberAuthDto);
-        return cartDao.save(new CartEntity(member.getId(), product.getId()));
-    }
-
-    private ProductEntity getProduct(final Long productId) {
-        return productDao.findById(productId)
-                .orElseThrow(() -> new IllegalArgumentException("존재하지 않는 상품입니다."));
+        final List<CartEntity> carts = cartDao.findAllByMemberId(member.getId());
+        return carts.stream()
+                .map(cart -> {
+                    final ProductEntity product = getProduct(cart.getProductId());
+                    return CartProductResponseDto.from(cart.getId(), product);
+                })
+                .collect(Collectors.toList());
     }
 
     private MemberEntity getMember(final MemberAuthDto memberAuthDto) {
@@ -42,18 +44,30 @@ public class CartService {
                 .orElseThrow(() -> new IllegalArgumentException("등록되지 않은 회원입니다."));
     }
 
-    public List<CartProductResponseDto> findCartItemsForMember(final MemberAuthDto memberAuthDto) {
-        final MemberEntity member = getMember(memberAuthDto);
-        final List<CartEntity> cartEntities = cartDao.findAllByMemberId(member.getId());
-        return cartEntities.stream()
-                .map(cartEntity -> {
-                    final ProductEntity product = getProduct(cartEntity.getProductId());
-                    return CartProductResponseDto.from(cartEntity.getId(), product);
-                })
-                .collect(Collectors.toList());
+    private ProductEntity getProduct(final Long productId) {
+        return productDao.findById(productId)
+                .orElseThrow(() -> new IllegalArgumentException("등록되지 않은 상품입니다."));
     }
 
+    @Transactional
+    public Long putInCart(final Long productId, final MemberAuthDto memberAuthDto) {
+        final MemberEntity member = getMember(memberAuthDto);
+        final ProductEntity product = getProduct(productId);
+        return cartDao.save(new CartEntity(member.getId(), product.getId()));
+    }
+
+    @Transactional
     public void removeCartItem(final Long cartId, final MemberAuthDto memberAuthDto) {
+        final MemberEntity member = getMember(memberAuthDto);
+        final CartEntity cart = getCart(cartId);
+        if (!cart.isOwner(member)) {
+            throw new IllegalArgumentException("장바구니 상품 소유자가 아닙니다.");
+        }
         cartDao.delete(cartId);
+    }
+
+    private CartEntity getCart(final Long cartId) {
+        return cartDao.findById(cartId)
+                .orElseThrow(() -> new IllegalArgumentException("등록되지 않은 장바구니 상품입니다."));
     }
 }
